salty-tundra-15690/app/requirements.txt
async-timeout
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography
Django
django-redis==5.2.0
numpy==1.24.3
oauthlib==3.2.2
pandas==2.0.1
plotly==5.14.1
pycparser==2.21
PyGithub==1.58.0
PyJWT==2.7.0
pypi==2.1
python-dateutil
python3-openid
pytz==2023.3
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse
whitenoise==6.4.0
gunicorn
dj-database-url
python-dotenv

GitPython

Flask
pika
SQLAlchemy
psycopg2-binary
gitpython
bandit
Django>=3.2,<4.0
pika
SQLAlchemy
psycopg2-binary
gitpython
bandit
celery
django-celery-results
celery
